mcp-servers/captcha-solver/pyproject.toml
[project]
name = "captcha-solver-mcp"
version = "2.0.0"
description = "Multi-strategy CAPTCHA solver MCP server — CLIP + VLM + External API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["mcp", "captcha", "hcaptcha", "recaptcha", "turnstile", "clip", "vlm", "vision", "ai", "solver"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "pillow>=10.0.0",
    "httpx>=0.27.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
vlm = ["google-genai>=1.0.0", "anthropic>=0.40.0", "openai>=1.0.0"]
browser = ["playwright>=1.40.0", "playwright-stealth>=2.0.0"]
api = ["fastapi>=0.110.0", "uvicorn>=0.29.0"]
all = ["captcha-solver-mcp[vlm,browser,api]"]

[project.scripts]
captcha-solver-mcp = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
